Fix a bug cause by the lack of TIMESTAMP function in sqlite
When adding an event calendar to the sidebar to the blog and clicking on
"all events", dotclear ends up in error with the message

near "'YYY-MM-DD HH:MM:DD'": syntax error (1)

with the date displayed the current date and time.
This is actually caused by an sql statement containing the TIMESTAMP
function, which sqlite does not recognize. This commit replaces all
occurrences of TIMESTAMP with nothing when the database connection is with a
sqlite database.